requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==4.1.0
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
attrs==23.2.0
awscli==1.32.22
blinker==1.7.0
boto3==1.34.22
botocore==1.34.22
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
docker==7.0.0
docutils==0.16
entrypoints==0.4
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.39.0
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
hub-sdk==0.0.3
idna==3.6
importlib-metadata==6.11.0
jaraco.classes==3.3.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keras==2.15.0
keyring==8.7
keyrings.alt==5.0.0
kiwisolver==1.4.5
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
openai==1.6.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
Pillow==9.2.0
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.7
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.7.2
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
scipy==1.11.4
seaborn==0.13.1
semver==3.0.2
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
streamlit==1.33.0
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ultralytics==8.1.1
urllib3==2.0.7
validators==0.22.0
vega-datasets==0.9.0
voluptuous==0.14.1
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
yolo==0.3.1
zipp==3.17.0
